Refactor models and parsers to standardize image_url and release_date handling
